refactor: 애플리케이션을 R2DBC가 아닌 TODO로 변경함
앞으로 R2DBC 외에 다른 기능들도 추가할 것이기 때문에 R2DBC라는 애플리케이션명은 옳지 않다고 판단함.
