--- src/main/java/com/r2dbc/R2DBCApplication.java
@@ -1,12 +1,0 @@
-package com.r2dbc;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class R2DBCApplication {
-
-    public static void main(String[] args) {
-        SpringApplication.run(R2DBCApplication.class, args);
-    }
-}
