--- src/main/java/com/r2dbc/jwt/JwtAuthenticationFilter.java
@@ -1,69 +1,0 @@
-package com.r2dbc.jwt;
-
-import lombok.RequiredArgsConstructor;
-import lombok.extern.slf4j.Slf4j;
-import org.springframework.http.HttpHeaders;
-import org.springframework.http.server.reactive.ServerHttpRequest;
-import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
-import org.springframework.security.core.Authentication;
-import org.springframework.security.core.GrantedAuthority;
-import org.springframework.security.core.authority.SimpleGrantedAuthority;
-import org.springframework.security.core.context.ReactiveSecurityContextHolder;
-import org.springframework.security.core.userdetails.User;
-import org.springframework.security.core.userdetails.UserDetails;
-import org.springframework.util.StringUtils;
-import org.springframework.web.server.ServerWebExchange;
-import org.springframework.web.server.WebFilter;
-import org.springframework.web.server.WebFilterChain;
-import reactor.core.publisher.Mono;
-
-import java.util.List;
-import java.util.stream.Collectors;
-
-@Slf4j
-@RequiredArgsConstructor
-public class JwtAuthenticationFilter implements WebFilter {
-
-    private final JwtTokenProvider tokenProvider;
-
-    @Override
-    public Mono<Void> filter(ServerWebExchange exchange, WebFilterChain chain) {
-        String token = resolveToken(exchange.getRequest());
-        if (token != null && tokenProvider.validateToken(token)) {
-            // 토큰이 유효하면 인증 정보 생성
-            String username = tokenProvider.getUsernameFromToken(token);
-
-            List<String> roles = tokenProvider.getRolesFromToken(token); // 1. roles 클레임 추출
-            List<GrantedAuthority> authorities = roles.stream()
-                    .map(SimpleGrantedAuthority::new)
-                    .collect(Collectors.toList());
-            authorities.forEach(System.out::println);
-
-            UserDetails userDetails = new User(
-                    username, // username
-                    "",       // password (JWT 인증에서는 필요 없으므로 빈 문자열)
-                    authorities // authorities
-            );
-
-            Authentication authentication = new UsernamePasswordAuthenticationToken(
-                    userDetails,
-                    null,
-                    authorities
-            );
-
-            // SecurityContext에 인증 정보 저장
-            return chain.filter(exchange).contextWrite(
-                ReactiveSecurityContextHolder.withAuthentication(authentication));
-        }
-        return chain.filter(exchange);
-    }
-    
-    // HTTP 요청 헤더에서 JWT 토큰 추출
-    private String resolveToken(ServerHttpRequest request) {
-        String bearerToken = request.getHeaders().getFirst(HttpHeaders.AUTHORIZATION);
-        if (StringUtils.hasText(bearerToken) && bearerToken.startsWith("Bearer ")) {
-            return bearerToken.substring(7);
-        }
-        return null;
-    }
-}
