--- src/main/java/com/r2dbc/post/dto/reponse/PostResponse.java
@@ -1,4 +1,0 @@
-package com.r2dbc.post.dto.reponse;
-
-public class PostResponse {
-}
