--- src/main/java/com/r2dbc/post/dto/request/PostCreateRequest.java
@@ -1,4 +1,0 @@
-package com.r2dbc.post.dto.request;
-
-public record PostCreateRequest(String title, String content) {
-}
