--- src/main/java/com/r2dbc/post/dto/request/PostUpdateRequest.java
@@ -1,4 +1,0 @@
-package com.r2dbc.post.dto.request;
-
-public record PostUpdateRequest(String title, String content) {
-}
