--- src/main/java/com/r2dbc/user/dto/request/UserLoginRequest.java
@@ -1,4 +1,0 @@
-package com.r2dbc.user.dto.request;
-
-public record UserLoginRequest(String username, String password) {
-}
