--- src/main/java/com/r2dbc/user/dto/request/UserUpdateRequest.java
@@ -1,4 +1,0 @@
-package com.r2dbc.user.dto.request;
-
-public record UserUpdateRequest(String name, String email) {
-}
