--- src/main/java/com/r2dbc/user/repository/UserRepository.java
@@ -1,10 +1,0 @@
-package com.r2dbc.user.repository;
-
-import com.r2dbc.user.domain.User;
-import org.springframework.data.repository.reactive.ReactiveCrudRepository;
-import reactor.core.publisher.Mono;
-
-public interface UserRepository extends ReactiveCrudRepository<User, Long> {
-    Mono<User> findByEmail(String email);
-    Mono<User> findByUsername(String username);
-}
