+++ src/main/java/com/todo/TODOApplication.java
@@ -1,0 +1,12 @@
+package com.todo;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class TODOApplication {
+
+    public static void main(String[] args) {
+        SpringApplication.run(TODOApplication.class, args);
+    }
+}
