+++ src/main/java/com/todo/exception/CustomException.java
@@ -1,0 +1,14 @@
+package com.todo.exception;
+
+import com.todo.exception.dto.ErrorMessage;
+import lombok.Getter;
+
+@Getter
+public class CustomException extends RuntimeException {
+    private final ErrorMessage errorMessage;
+
+    public CustomException(ErrorMessage errorMessage, String message) {
+        super(message);
+        this.errorMessage = errorMessage;
+    }
+}
