+++ src/main/java/com/todo/post/dto/reponse/PostResponse.java
@@ -1,0 +1,4 @@
+package com.todo.post.dto.reponse;
+
+public class PostResponse {
+}
