+++ src/main/java/com/todo/post/dto/reponse/UserWithPostsResponse.java
@@ -1,0 +1,11 @@
+package com.todo.post.dto.reponse;
+
+import com.todo.post.domain.Post;
+import com.todo.user.domain.User;
+
+import java.util.List;
+
+public record UserWithPostsResponse(
+        User user,
+        List<Post> posts
+) {}
