+++ src/main/java/com/todo/post/dto/request/PostCreateRequest.java
@@ -1,0 +1,4 @@
+package com.todo.post.dto.request;
+
+public record PostCreateRequest(String title, String content) {
+}
