+++ src/main/java/com/todo/post/dto/request/PostUpdateRequest.java
@@ -1,0 +1,4 @@
+package com.todo.post.dto.request;
+
+public record PostUpdateRequest(String title, String content) {
+}
