+++ src/main/java/com/todo/user/dto/request/UserLoginRequest.java
@@ -1,0 +1,4 @@
+package com.todo.user.dto.request;
+
+public record UserLoginRequest(String username, String password) {
+}
