+++ src/main/java/com/todo/user/dto/request/UserUpdateRequest.java
@@ -1,0 +1,4 @@
+package com.todo.user.dto.request;
+
+public record UserUpdateRequest(String name, String email) {
+}
