+++ src/main/java/com/todo/user/repository/UserRepository.java
@@ -1,0 +1,10 @@
+package com.todo.user.repository;
+
+import com.todo.user.domain.User;
+import org.springframework.data.repository.reactive.ReactiveCrudRepository;
+import reactor.core.publisher.Mono;
+
+public interface UserRepository extends ReactiveCrudRepository<User, Long> {
+    Mono<User> findByEmail(String email);
+    Mono<User> findByUsername(String username);
+}
