--- src/test/java/com/r2dbc/R2DBCApplicationTests.java
@@ -1,13 +1,0 @@
-package com.r2dbc;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class R2DBCApplicationTests {
-
-    @Test
-    void contextLoads() {
-    }
-
-}
