+++ src/test/java/com/todo/TODOApplicationTests.java
@@ -1,0 +1,13 @@
+package com.todo;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class TODOApplicationTests {
+
+    @Test
+    void contextLoads() {
+    }
+
+}
